Minor code change in ConnectedDevicesViewModel
We're chasing something, and we haven't caught it yet. In the meantime, better to commit the small improvements we're making and that should not be affecting the outcome of what we're looking for, so that if we need to wipe the slate clean we don't lose them and have to do them again.