Improved Macro Selection Animation
Basically, we break "less" the SwiftUI View Hierarchy. So no jumping. And instead, if the selected Macro is "None", then the buttons are disabled. Makes sense in a way. Surprisingly it works better in the VStack below the List than as a Section in the List, which we tested to see what is tripping up the SwiftUI Animations.