Change: Editing Macro Commands is now done in its own Screen
We can't do it in an Alert / Dialog, at least not a native one, because it's too complex a process for what Apple allows us to do inside a native Alert. So we need to make an Edit Screen just for that. Which opens up design freedom for us, at the expense of more navigational complexity for the user.