Amended Item #17: Changes to Inspector "Nav Bar"
Mostly we changed the color. If we put a NavBar in the "main Inspector" screen, when opening Attribute Table it just adds another NavBar below it, which we don't want. It either works this way, which I doubt Filip is going to like, or we don't use it as an Inspector but as a regular "sheet " instead. This is the most likely outcome I think.