Temporarily removed Macros from UART
When we can implement them properly, we will. But for now, let's not. Better to remove it than risk keeping it in, and that it slips into release. Specially since we're having to check so many different things and we can't just seem to be able to come back to UART at the moment.