Integrated ScannerViewModel code into ConnectedDevicesViewModel
We could not keep it separate because of the @Published -> @Binding -> Binding parameter. I could not figure it out, and my brain wasn't up for it. So instead of uhm, being frustrated. I just "solved it" by putting all the code in one place. Maybe I'm not making the best decisions ever, but I just want to keep going and not enter the weekend with a bit of frustration.