Deleted HUDState
Now that I've removed the code, I find it funny that I understand what Nick was going for. But anyway, we can always bring it back if we need to. But for now we can keep just watching the logs. Though this was a very Swifty idea. Or SwiftUI idea.