Deleted SensorSettings and SensorSettingsVM
Again, they were not being used? I assumed this was the code for the "Sensor Calibration Setup" but... no. I might need them again, but at the moment I'm trying to find my bearings so, best to keep just what works and is useful currently. If something's missing we can always bring it back. That's what git is for.